# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.TransferRequestSwag:
    properties:
      jumlah_transfer:
        type: integer
      pin:
        type: string
      tujuan_transfer:
        type: string
    type: object
...
  dto.WithdrawDto:
    properties:
      withdraw:
        type: integer
    type: object
...
  model.Transfer:
    properties:
      id:
        type: string
      jenis_transfer:
        type: string
      jumlah_transfer:
        type: integer
      nama_penerima:
        type: string
      nama_pengirim:
        type: string
      trx_id:
        type: string
      tujuan_transfer:
        type: string
      user_id:
        type: string
    type: object
...
  model.Withdraw:
    properties:
      created_at:
        type: string
      id:
        type: string
      user_id:
        type: string
      withdraw:
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: A tag service API in Go using Gin framework
  title: Tag Service API
  version: "1.0"
paths:
...
  /transfer:
    post:
      consumes:
      - application/json
      description: User may transfer to other user
      parameters:
      - description: Bearer
        in: header
        name: Authorization
        required: true
        type: string
      - description: Payload data
        in: body
        name: Data
        required: true
        schema:
          $ref: '#/definitions/dto.TransferRequestSwag'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Transfer'
      security:
      - ApiKeyAuth: []
      summary: Create request Transfer
      tags:
      - Transfer
  /transfer/history/admin/receive/{id}:
    get:
      consumes:
      - application/json
      description: Admin can get user histories receive transfer @3datas/page
      parameters:
      - description: Bearer
        in: header
        name: Authorization
        required: true
        type: string
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Transfer'
            type: array
      security:
      - ApiKeyAuth: []
      summary: Admin can get user histories receive transfer @3datas/page
      tags:
      - Admin
  /transfer/history/admin/send/{id}:
    get:
      consumes:
      - application/json
      description: Admin can get their histories send transfer @3datas/page
      parameters:
      - description: Bearer
        in: header
        name: Authorization
        required: true
        type: string
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Transfer'
            type: array
      security:
      - ApiKeyAuth: []
      summary: Admin can get their histories send transfer @3datas/page
      tags:
      - Admin
  /transfer/history/receive:
    get:
      consumes:
      - application/json
      description: User can get their histories receive transfer @3datas/page
      parameters:
      - description: Bearer
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Transfer'
            type: array
      security:
      - ApiKeyAuth: []
      summary: User can get their histories receive transfer @3datas/page
      tags:
      - Transfer
  /transfer/history/send:
    get:
      consumes:
      - application/json
      description: User can get their histories send transfer @3datas/page
      parameters:
      - description: Bearer
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Transfer'
            type: array
      security:
      - ApiKeyAuth: []
      summary: User can get their histories send transfer @3datas/page
      tags:
      - Transfer
  /transfer/withdraw:
    get:
      consumes:
      - application/json
      description: User can get histories withdraw
      parameters:
      - description: Bearer
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Withdraw'
            type: array
      security:
      - ApiKeyAuth: []
      summary: User can get histories withdraw
      tags:
      - Transfer
    post:
      consumes:
      - application/json
      description: User can withdraw their balance to rekening bank
      parameters:
      - description: Bearer
        in: header
        name: Authorization
        required: true
        type: string
      - description: Withdraw ammount
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.WithdrawDto'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Withdraw'
      security:
      - ApiKeyAuth: []
      summary: User can withdraw their balance to rekening bank
      tags:
      - Transfer
...
swagger: "2.0"